scripts/requirements.txt
pyyaml>=6.0
kubernetes>=28.1.0 